+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Explicit support for plain/no parse mode
+
+### Fixed
+
+- Documentation issues, related to parse modes
+
 ## [1.2.0] - 2025.11.04
 
 ### Added